week10/day5/project_example/Venandi-master/requirements.txt
flask
python-dotenv
dnspython
pandas
sklearn
requests
bleach
selenium
